pyproject.toml
[project]
name = "fast-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "paho-mqtt>=2.1.0",
    "ruff>=0.11.2",
]
